chore: remove unused Timeline class, SPECTRA_MULTIPROC macro, and obsolete build artifacts
- Remove Timeline class and timeline.hpp header as animation system now uses TimelineEditor/TransitionEngine
- Remove Animator::add_timeline()/remove_timeline() methods and timelines_ member as they depended on removed Timeline
- Remove SPECTRA_MULTIPROC compile definition from CMakeLists.txt as runtime mode is now determined at runtime via AppConfig::socket_path
- Remove conditional compilation of app_in

+++ src/ui/session_runtime.cpp
@@ -459,9 +459,9 @@ FrameState SessionRuntime::tick(FrameScheduler& scheduler,
             }
 
             fprintf(stderr,
-                    "[move]   DONE: fig=%u moved %u→%u "
+                    "[move]   DONE: fig=%lu moved %u→%u "
                     "(src_figs=%zu dst_figs=%zu split=%d)\n",
-                    pm.figure_id,
+                    static_cast<unsigned long>(pm.figure_id),
                     src_wctx->id,
                     dst_wctx->id,
                     src_wctx->assigned_figures.size(),
